+++ include/arviss/common/instruction.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "arviss/types.h"
+#include "arviss/common/types.h"
 
 namespace arviss
 {
