+++ include/arviss/rv32/concepts.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "arviss/core/concepts.h"
-#include "arviss/types.h"
+#include "arviss/common/types.h"
 
 #include <concepts>
 #include <string>
