+++ include/arviss/rv32/dispatchers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "arviss/core/concepts.h"
-#include "arviss/instruction.h"
+#include "arviss/common/instruction.h"
 #include "arviss/rv32/concepts.h"
 
 namespace arviss
